types cleanup and prop rename
