+++ packages/thirdweb/src/react/web/wallets/in-app/InAppWalletSocialLogin.tsx
@@ -1,6 +1,6 @@
 "use client";
 import { useEffect, useRef, useState } from "react";
-import type { InAppWalletSocialAuth } from "../../../../wallets/in-app/core/wallet/index.js";
+import type { InAppWalletSocialAuth } from "../../../../wallets/in-app/core/wallet/types.js";
 import type { Wallet } from "../../../../wallets/interfaces/wallet.js";
 import { useConnectUI } from "../../../core/hooks/others/useWalletConnectionCtx.js";
 import { Spacer } from "../../ui/components/Spacer.js";
