+++ packages/thirdweb/src/react/web/wallets/in-app/openOauthSignInWindow.ts
@@ -1,4 +1,4 @@
-import type { InAppWalletSocialAuth } from "../../../../wallets/in-app/core/wallet/index.js";
+import type { InAppWalletSocialAuth } from "../../../../wallets/in-app/core/wallet/types.js";
 import type { Theme } from "../../ui/design-system/index.js";
 
 function getBodyTitle(authOption: InAppWalletSocialAuth) {
