+++ packages/thirdweb/src/wallets/in-app/core/wallet/types.ts
@@ -1,29 +1,36 @@
+import type { Chain } from "../../../../chains/types.js";
 import type { ThirdwebClient } from "../../../../client/client.js";
-import type { AsyncStorage } from "../../../storage/AsyncStorage.js";
+import type { SmartWalletOptions } from "../../../smart/types.js";
+import type {
+  MultiStepAuthArgsType,
+  SingleStepAuthArgsType,
+} from "../authentication/type.js";
 
-export type InAppWalletCreationOptions = {
-  /**
-   * A client is the entry point to the thirdweb SDK.
-   * It is required for all other actions.
-   * You can create a client using the `createThirdwebClient` function. Refer to the [Creating a Client](https://portal.thirdweb.com/typescript/v5/client) documentation for more information.
-   *
-   * You must provide a `clientId` or `secretKey` in order to initialize a client. Pass `clientId` if you want for client-side usage and `secretKey` for server-side usage.
-   *
-   * ```tsx
-   * import { createThirdwebClient } from "thirdweb";
-   *
-   * const client = createThirdwebClient({
-   *  clientId: "<your_client_id>",
-   * })
-   * ```
-   */
+export type InAppWalletConnectionOptions = (
+  | MultiStepAuthArgsType
+  | SingleStepAuthArgsType
+) & {
   client: ThirdwebClient;
-  /**
-   * Storage interface of type [`AsyncStorage`](https://portal.thirdweb.com/references/typescript/v5/AsyncStorage) to save connected wallet data to the storage for auto-connect.
-   * If not provided, no wallet data will be saved to the storage by thirdweb SDK
-   */
-  storage?: AsyncStorage;
+  chain?: Chain;
 };
+
+export type InAppWalletAutoConnectOptions = {
+  client: ThirdwebClient;
+  chain?: Chain;
+};
+
+export type InAppWalletSocialAuth = "google" | "apple" | "facebook";
+
+export type InAppWalletAuth = "email" | "phone" | InAppWalletSocialAuth;
+
+export type InAppWalletCreationOptions =
+  | {
+      auth?: {
+        options: InAppWalletAuth[];
+      };
+      smartAccount?: SmartWalletOptions;
+    }
+  | undefined;
 
 export type AuthenticatedUser = {
   email: string | undefined;
